requirements.txt
numpy==1.26.3
streamlit==1.31.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
groq==0.4.1